Move _global.scss and transition.utils.ts in common
