+++ libs/search-widget/src/Widget.svelte
@@ -20,7 +20,7 @@
   import type { KBStates, Resource } from '@nuclia/core';
   import { Observable } from 'rxjs';
   import { setupSuggestionsAndPredictions, setupTriggerSearch } from './core/search-bar';
-  import globalCss from './_video-widget/_global.scss';
+  import globalCss from './common/_global.scss';
 
   export let backend = 'https://nuclia.cloud/api';
   export let widgetid = '';
