+++ libs/search-widget/src/_video-widget/SearchBar.svelte
@@ -9,7 +9,7 @@
   import { setLang } from '../core/i18n';
   import SearchInput from '../widgets/search-input/SearchInput.svelte';
   import { setupTriggerSearch } from '../core/search-bar';
-  import globalCss from './_global.scss';
+  import globalCss from '../common/_global.scss';
 
   export let backend = 'https://nuclia.cloud/api';
   export let widgetid = '';
