+++ libs/search-widget/src/common/_global.scss
