+++ libs/search-widget/src/common/transition.utils.ts
